backend/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.21.0
psutil==5.9.8
